+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - (Feature) (Networking) ArangoRotue WebSocket Support
 - (Feature) (Scheduler) Helm Driver Param
 - (Feature) (Integration) Services Endpoint
+- (Feature) (Platform) Storage
 
 ## [1.2.43](https://github.com/arangodb/kube-arangodb/tree/1.2.43) (2024-10-14)
 - (Feature) ArangoRoute CRD